production_agentic_pipeline/requirements.txt
streamlit
pandas
pyodbc
requests
python-dotenv
numpy
plotly
openpyxl
xlrd
